+++ source/index.html
@@ -1123,11 +1123,6 @@
       <div class="screen__wrapper">
         <div class="result result--trip">
           <div class="result__image">
-            <!-- <picture>
-              <source srcset="img/result1-mob.png 1x, img/result1-mob@2x.png 2x"
-                media="(max-width: 768px) and (orientation: portrait)">
-              <img src="img/result1.png" srcset="img/result1@2x.png 2x" alt="">
-            </picture> -->
             <canvas id="sea-calf-canvas"></canvas>
           </div>
           <h2 class="result__title">
@@ -1308,11 +1303,7 @@
       <div class="screen__wrapper">
         <div class="result result--negative">
           <div class="result__image">
-            <picture>
-              <source srcset="img/result3-mob.png 1x, img/result3-mob@2x.png 2x"
-                media="(max-width: 768px) and (orientation: portrait)">
-              <img src="img/result3.png" srcset="img/result3@2x.png 2x" alt="">
-            </picture>
+            <canvas id="crocodile"></canvas>
           </div>
           <h2 class="result__title">
             <span class="visually-hidden">Не угадал!</span>
